load_tests/requirements.txt
# Load testing dependencies
locust>=2.20.0
python-jose>=3.3.0
